requirements.txt
# TESLA TPC Streamlit Analytics Platform
# Python Package Dependencies

# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database Connectivity
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.8.0

# Visualization
plotly>=5.15.0
altair>=5.0.0

# Date/Time Handling
python-dateutil>=2.8.0

# Configuration & Secrets
toml>=0.10.2

# Development & Testing (Optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Performance (Optional but recommended)
watchdog>=3.0.0

pathlib2>=2.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
toml>=0.10.0                